Protect org.apache.http.legacy API
Add build targets to check the API and update the API along with API
description files to describe the state of the API.

Tested by invoking make checkapi to make sure that it detected the API
change and failed the build. Then invoke make update-api to update the
API and then make checkapi again to ensure that this time the build
succeeded.

Bug: 18027885
Test: see above
Change-Id: Id958c5871cd43773f5f3b56687e9b5fab023c6f0