Convert use of Time.format() to an alternative
Convert use of Time.format() to an alternative.
This avoids future Y2038 issues associated with Time.

The code has been changed to use GregorianCalendar
which has lenient rounding behavior when values are
out of range (similar to android.text.format.Time).
This is unlike LocalDateTime, which throws exceptions.

There are no tests for this code.

Bug: 16550209
Test: build only
Change-Id: I43e016e7ed80f81887708c4a0f0123ce4baaf4b8