Fix ReferencesHidden issues in apache http
We had RequestQueue had several public APIs that used
hidden Connection and Request classes. This made these
impossible to call for developers and in fact stubs would
not even compile without a metalava workaround for this.

This CL remove the disabling of ReferencesHidden as ignoring
ReferencesHidden leads to pulling in hidden classes
into API surfaces which is always a bug. Removing hiding here
so this option to suppress can be removed completely.

Test: make -j -> passes
Change-Id: I8e02af533be5cd0eccad0d9eed0a85c9b1b9b3e8